+++ Cine.Core/Sala.cs
@@ -4,8 +4,8 @@ namespace Cine.Core
 {
     public class Sala
     {
-        public int idsala { get; set;}
-        public int piso { get; set;}
+        public sbyte idsala { get; set;}
+        public sbyte piso { get; set;}
         public ushort capacidad { get; set;}
     }
